+++ src/actions/test/setBlockTimestampInterval.ts
@@ -11,7 +11,7 @@ export type SetBlockTimestampIntervalParameters = {
 }
 
 /**
- * Similar to [`increaseTime`](https://viem.sh/docs/actions/test/increaseTime), but sets a block timestamp `interval`. The timestamp of future blocks will be computed as `lastBlock_timestamp` + `interval`.
+ * Similar to [`increaseTime`](https://viem.sh/docs/actions/test/increaseTime.html), but sets a block timestamp `interval`. The timestamp of future blocks will be computed as `lastBlock_timestamp` + `interval`.
  *
  * - Docs: https://viem.sh/docs/actions/test/setBlockTimestampInterval.html
  *